move files, fix tests.
